requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
elasticsearch==7.10.1
fastapi==0.63.0
h11==0.12.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
numpy==1.19.5
pandas==1.2.0
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.10.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
query==0.1.4
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
starlette==0.13.6
tqdm==4.54.1
Twisted==20.3.0
urllib3==1.26.2
uvicorn==0.13.3
w3lib==1.22.0
zope.interface==5.2.0
